pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forge"
version = "0.1.0"
description = "Feedback-Oriented Reward & Gaussian Estimation"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pandas",
    "scikit-learn",
    "mord",
    "gpytorch",
    "torch",
    "scipy",
    "pyyaml",
    "click",
    "rich",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
]

[project.scripts]
forge = "forge.cli:cli"

[tool.setuptools.packages.find]
include = ["forge*"]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
